refactor: consolidate codemode tests, improve error handling, and add python-dotenv support
