+++ pyproject.toml
@@ -17,6 +17,7 @@ dependencies = [
     "cryptography>=45.0.4",
     "structlog>=25.4.0",
     "questionary>=2.1.0",
+    "python-dotenv>=1.1.1",
 ]
 
 [project.optional-dependencies]
